+++ src/main/java/roomescape/error/exception/ThemeNotExistsException.java
@@ -1,0 +1,8 @@
+package roomescape.error.exception;
+
+public class ThemeNotExistsException extends NotExistsException {
+
+    public ThemeNotExistsException() {
+        super("해당 테마가");
+    }
+}
