--- src/main/java/roomescape/exception/ThemeNotExistsException.java
@@ -1,8 +1,0 @@
-package roomescape.exception;
-
-public class ThemeNotExistsException extends NotExistsException{
-
-    public ThemeNotExistsException() {
-        super("해당 테마가");
-    }
-}
