+++ src/main/java/roomescape/reservation/service/ReservationRepository.java
@@ -1,10 +1,11 @@
-package roomescape.reservation;
+package roomescape.reservation.service;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.reservation.Reservation;
 import roomescape.reservationTime.ReservationTime;
 import roomescape.theme.Theme;
